requirements.txt
discord.py==2.4.0
motor==3.4.0
pymongo==4.7.0
python-dotenv==1.0.1
dnspython==2.6.1